requirements.txt
emoji==0.6.0
python-telegram-bot==21.6
requests
rapidfuzz
Pillow
